// src/app/shared/interfaces/user-card-interface.ts
export interface UserCardInterface {
  id: string;
  name: string;
  location: string;
  institution: string;
  route?: string;
  phone?: string;
  email?: string;
  member_number?: string;
  sub_county_name?: string;
  county_name?: string;
}
export interface UserCardInterface {
  id: string;
  name: string;
  location: string;
  institution: string;
  route?: string;
  phone?: string;
  email?: string;
  member_number?: string;
  sub_county_name?: string;
  county_name?: string;
}
export interface FarmerUserCardInterface {
  route: string;
  phone: string;
  email: string;
}